Moved the last bits over and deleted old class
